refactred code
