+++ FitnessApp/Home/Models/Workout.swift
@@ -1,0 +1,18 @@
+//
+//  Workout.swift
+//  FitnessApp
+//
+//  Created by Ricardo Garza on 1/8/25.
+//
+
+import SwiftUI
+
+struct Workout {
+    let id: Int
+    let title: String
+    let image: String
+    let tintColor: Color
+    let duration : String
+    let date : String
+    let calories : String
+}
