+++ FitnessApp/Home/Views/ActivityCard.swift
@@ -5,18 +5,7 @@
 //  Created by Ricardo Garza on 1/7/25.
 //
 
-import SwiftUI
-
-struct Activity {
-    let id: Int
-    let title: String
-    let subtitle: String
-    let image: String
-    let tintColor: Color
-    let amount: String
-}
-
- 
+import SwiftUI 
 
 struct ActivityCard: View {
     @State var activity: Activity
